global-hotspots/apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
cryptography==44.0.0
email-validator==2.2.0
httpx==0.28.1
feedparser==6.0.11
